update anet.c to redis-2.8.19 and simplify Quickstart
The Quickstart part is more like how to run kids in production, however,
without package, it is clumsy to deploy kids.

It is better to have a configuration file which listens on localhost:3388,
and stores message to `/tmp` to make it easier to play with kids locally,
just like redis-server.

Only after we have release a deb package can we add the instructions about
how to run kids in production.